+++ app/src/main/java/com/example/huaheejqc/data/Chat.kt
@@ -1,3 +1,3 @@
 package com.example.huaheejqc.data
 
-data class Chat(var id: String, var lastMsg: String, var timestamp: Int)
+data class Chat(var id: String, var lastMsg: String, var timestamp: Long)
